add delete e getById

+++ listaDesejo/src/main/java/com/flaviadev/listaDesejo/repository/DesejoRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.flaviadev.listaDesejo.model.Desejo;
 
-public interface DesejoRepository  extends JpaRepository<Desejo, Integer>{
+public interface DesejoRepository  extends JpaRepository<Desejo, Long>{
 
 }
